+++ Gymopedia.Domain/Models/ClientToCoach.cs
@@ -4,7 +4,7 @@ namespace Gymopedia.Domain.Models
     public class ClientToCoach
     {
         public int Id { get; set; }
-        public int CoachId { get; set; }
-        public int ClientId { get; set; }
+        public long CoachId { get; set; }
+        public long ClientId { get; set; }
     }
 }
